requirements.txt
asgiref==3.4.1
Django==2.1.5
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.2.0
